Feat : spring security for login and create member entity
we use spring security that is under framework in spring framework
this use session storage.

member entity is a intormation that our notice board consumer
we use entity name 'member' because user class is exist in spring secutity and we need to use that also for login.
so their naming is need to different.